fix wget in shell scripts for 3D to wget only if image files don't already exist

+++ demos/3D/demo3D11/demo3D11.sh
@@ -41,13 +41,19 @@
 # For additional information visit http://murphylab.web.cmu.edu or
 # send email to murphy@cmu.edu
 
-wget -nc --quiet http://www.cellorganizer.org/Downloads/v2.7/docker/demo3D11.tgz
+directory=../../../images/HeLa/2D/LAM
 
-mkdir -p ../../../images/HeLa/3D
+if [ ! -d "$directory" ]; then
 
-tar -xvf demo3D11.tgz -C ../../../images/HeLa/3D/
+	wget -nc --quiet http://www.cellorganizer.org/Downloads/v2.7/docker/demo3D11.tgz
 
-rm -f demo3D11.tgz
+	mkdir -p ../../../images/HeLa/3D/processed
+
+	tar -xvf demo3D11.tgz -C ../../../images/HeLa/3D/processed/
+
+	rm -f demo3D11.tgz
+
+fi
 
 echo -e "options.sampling.method = 'disc';
 options.debug = true;
